+++ KeireCore/Include/KeireInternal/UiInputInternal.h
@@ -6,5 +6,6 @@ namespace Keire::Detail
 {
     [[nodiscard]] bool UiBackendKeyDown(UiKey key) noexcept;
     [[nodiscard]] bool UiBackendKeyPressed(UiKey key) noexcept;
+    void UiBackendRequestTextInput() noexcept;
     [[nodiscard]] std::string UiBackendTextInput();
 } // namespace Keire::Detail